Tproxy Cluster

+++ tdxctl/src/fde_setup.rs
@@ -79,6 +79,8 @@ struct InstanceInfo {
     #[serde(default)]
     bootstrapped: Option<bool>,
     #[serde(with = "hex_bytes", default)]
+    instance_id_seed: Vec<u8>,
+    #[serde(with = "hex_bytes", default)]
     instance_id: Vec<u8>,
     #[serde(with = "hex_bytes", default)]
     app_id: Vec<u8>,
@@ -86,7 +88,8 @@ struct InstanceInfo {
 
 impl InstanceInfo {
     fn is_bootstrapped(&self) -> bool {
-        self.bootstrapped.unwrap_or(!self.instance_id.is_empty())
+        self.bootstrapped
+            .unwrap_or(!self.instance_id_seed.is_empty())
     }
 }
 
@@ -340,11 +343,17 @@ impl SetupFdeArgs {
         let existing_rootfs_hash = fs::read(&hash_file).unwrap_or_default();
         if existing_rootfs_hash != self.rootfs_hash {
             info!("Rootfs hash changed, upgrading the rootfs");
+            let upgrading_file = self.rootfs_dir.join(".rootfs_upgrading");
+            if upgrading_file.exists() {
+                bail!("The previous rootfs upgrade is not finished, aborting");
+            }
+            fs::write(&upgrading_file, "").context("Failed to touch upgrading file")?;
             if hash_file.exists() {
                 fs::remove_file(&hash_file).context("Failed to remove old rootfs hash file")?;
             }
             host.notify_q("boot.progress", "upgrading rootfs").await;
             self.extract_rootfs(&self.rootfs_hash).await?;
+            fs::remove_file(&upgrading_file).context("Failed to remove upgrading file")?;
         }
         Ok(())
     }
@@ -478,14 +487,19 @@ impl SetupFdeArgs {
         }
 
         let disk_reusable = (!key_provider.is_none()) || !self.rootfs_encryption;
-        if (!disk_reusable) || instance_info.instance_id.is_empty() {
-            instance_info.instance_id = {
+        if (!disk_reusable) || instance_info.instance_id_seed.is_empty() {
+            instance_info.instance_id_seed = {
                 let mut rand_id = vec![0u8; 20];
                 getrandom::getrandom(&mut rand_id)?;
-                rand_id.extend_from_slice(&instance_info.app_id);
-                sha256(&rand_id)[..20].to_vec()
+                rand_id
             };
         }
+        let instance_id = {
+            let mut id_path = instance_info.instance_id_seed.clone();
+            id_path.extend_from_slice(&instance_info.app_id);
+            sha256(&id_path)[..20].to_vec()
+        };
+        instance_info.instance_id = instance_id.clone();
         if !kms_enabled && instance_info.app_id != truncated_compose_hash {
             bail!("App upgrade is not supported without KMS");
         }
@@ -496,7 +510,7 @@ impl SetupFdeArgs {
         extend_rtmr3("rootfs-hash", &self.rootfs_hash)?;
         extend_rtmr3("app-id", &instance_info.app_id)?;
         extend_rtmr3("compose-hash", &compose_hash)?;
-        extend_rtmr3("instance-id", &instance_info.instance_id)?;
+        extend_rtmr3("instance-id", &instance_id)?;
         extend_rtmr3("boot-mr-done", &[])?;
         // Show the RTMR
         cmd_show()?;
